API: removed SolverForwardEuler since we can now create a SolverOde object rather than a Solver object.
