+++ src/api/libopencor/module/libopencor
@@ -20,7 +20,6 @@ limitations under the License.
 #include "libopencor/issue.h"
 #include "libopencor/logger.h"
 #include "libopencor/solver.h"
-#include "libopencor/solverforwardeuler.h"
 #include "libopencor/solverinfo.h"
 #include "libopencor/solverode.h"
 #include "libopencor/solverproperty.h"
